Move openapi util from kubectl to cli-runtime
